CompositeStatusService actions and reducers typed
